Unused UI Removed + Queue Modifications
The functions to create UI Bars of the Tower and the Barracks are comented since they are not used and the unsused UI Creation Bars have been erased.
Meanwhile, the queue to create units will now take into cosideration the edge collected and the creation bars won't show up if the player has less edge that the creation cost for every unit.